pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coco2yolo-kaggle"
version = "0.1.0"
description = "Convert COCO dataset to YOLO format in Kaggle environment"
readme = "README.md"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.6"
dependencies = [
    "numpy",
    "tqdm",
    "pillow",
    "pandas",
]

[project.urls]
Homepage = "https://github.com/yourusername/coco2yolo-kaggle"
Issues = "https://github.com/yourusername/coco2yolo-kaggle/issues"

[project.scripts]
coco2yolo-kaggle = "coco2yolo_kaggle.cli:main"